[WIP] Gather: Use utility function to append to debug logs
Using echo and then piping it to a log file is a pattern
that can be seen throughout this project. Doing so requires
quite a log of syntax.

The idea is to have a file with common utility functions and source
them at the beginning of the run.

We already pass `BASE_COLLECTION_PATH` as an argument to every script
so this approach is just building on top of that.

Another fix included in this patch is for out of scope variable reference
at preinstall.sh:29. Shell scripts do not carry over variable definitions.

Signed-off-by: black-dragon74 <niryadav@redhat.com>