Use ENV to determine the interpreter path
This approach is preferred on systems where the path
of the interpreter might not be the one we defined.

We do not use interpreter args anywhere in this project and hence
this patch can be applied as a drop-in replacement. This makes the project
more portable and in sync with OpenShift mg.

Signed-off-by: black-dragon74 <niryadav@redhat.com>