Improve handling of BASE_COLLECTION_PATH
And its resolution in case it is not defined. We never call any of the
collection scripts without them expecting basepath unless being called
explicitly from the container.

In that case, it will resolve to PWD and the script will execute as
expected without any overlaps as exported shell variables have a TTL
until they are subprocesses of the current shell.

Signed-off-by: black-dragon74 <niryadav@redhat.com>