Include pod metadata in must-gather output
This PR adds support of fetching the pod metadata
using the kubernetes API.

An additional pod-metadata.json file is created in the
report with the dump of entire metadata for the pod.

A much simpler debug log stating the image version being
used is printed to the console as well.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>